+++ vector/src/main/java/im/vector/app/features/userdirectory/UserDirectoryViewState.kt
@@ -31,7 +31,7 @@ data class UserDirectoryViewState(
         val createAndInviteState: Async<String> = Uninitialized,
         val directorySearchTerm: String = "",
         val filterKnownUsersValue: Option<String> = Option.empty(),
-        val isThereAnExistingRoom: Boolean = false
+        val existingDmRoomId: String? = null
 ) : MvRxState {
 
     constructor(args: KnownUsersFragmentArgs) : this(excludedUserIds = args.excludedUserIds)